add google login to welcome screen
